third_party/WebKit/LayoutTests/http/tests/history/client-redirect-after-push-state.html

<!DOCTYPE html>
<html>
<head>
  <script src="../../js-test-resources/js-test.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description('Test that history.pushState can be called between scheduling and firing of a client redirect and still create just one new history entry. This was regressed in r149960');

if (window.testRunner)
    testRunner.dumpChildFrames(true);

function run()
{
    var iframe = document.createElement("iframe");
    document.body.appendChild(iframe);
    iframe.src = "resources/redirect-target.html#2";
    window.history.pushState("", "", location.href);
};

window.onload = function() {
    setTimeout(run, 0);
}

var jsTestIsAsync = true;
</script> 
</body>
</html>
