third_party/blink/web_tests/http/tests/history/client-redirect-after-push-state-expected.txt

Test that history.pushState can be called between scheduling and firing of a client redirect and still create just one new history entry. This was regressed in r149960

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
This page is the target of a redirect.

PASS: History item count should be 2 and is.


============== Back Forward List ==============
        http://127.0.0.1:8000/history/client-redirect-after-push-state.html
            about:blank (in frame "<!--framePath //<!--frame0-->-->")
curr->  http://127.0.0.1:8000/history/client-redirect-after-push-state.html
            http://127.0.0.1:8000/history/resources/redirect-target.html#2 (in frame "<!--framePath //<!--frame0-->-->")
===============================================
